FOCES pipeline: move BrightnessProfileFigure from reduce_doublefiber to common
